[~] bow get bonus dmg

+++ src/generator/load.c
@@ -73,7 +73,7 @@ game_t *load_gameb(void)
     read(fd, &game->speed, sizeof(int)) != sizeof(int) ? (game = 0) :
     read(fd, &game->attack, sizeof(int)) != sizeof(int) ? (game = 0) :
     read(fd, &game->nofall, sizeof(int)) != sizeof(int) ? (game = 0) :
-    read(fd, &game->skills, sizeof(int)) != sizeof(int) ? (game = 0) :
+    read(fd, &game->skills, sizeof(float)) != sizeof(int) ? (game = 0) :
     read(fd, &game->mob_kill, sizeof(int)) != sizeof(int) ? (game = 0) :
     0;
     close(fd);
